refactor(backend-service-game): move CardFixtures
